Close #106 - Move ExecutorServiceOps from test to main
