+++ cats-effect/src/test/scala/effectie/cats/CanCatchSpec.scala
@@ -6,7 +6,8 @@ import cats.effect._
 import cats.implicits._
 
 import effectie.Effectful._
-import effectie.{ExecutorServiceOps, SomeControlThrowable}
+import effectie.SomeControlThrowable
+import effectie.concurrent.ExecutorServiceOps
 
 import hedgehog._
 import hedgehog.runner._
